fix: make pkg client independent to other inner pkgs
